% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_wikipedia_page.R
\name{geocode_wikipedia_page}
\alias{geocode_wikipedia_page}
\title{Get's Wikipedia Page Coordinates}
\usage{
geocode_wikipedia_page(page_name)
}
\arguments{
\item{page_name}{Wikipedia page name. Should be a character}
}
\value{
A 2 dimensional list, where the first entry is latitude and the second is longitude
}
\description{
This function takes a wikipedia page name and returns it's location in latitude and longitide
}
