% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_wikipedia_page.R
\name{search_for_wikipedia_page}
\alias{search_for_wikipedia_page}
\title{Searches for a Wikipedia Page using a variety of methods}
\usage{
search_for_wikipedia_page(
  search_string,
  method = "rvest-google",
  use_first_result = TRUE
)
}
\arguments{
\item{search_string}{The string to be searched. Should be a character}

\item{method}{The method to be used to search}

\item{use_first_result}{Boolean variable. Default True. If True returns the first link, if false, returns a dataframe of all results}
}
\value{
Either a page_name in character form or a dataframe
}
\description{
This function takes a location string and finds it's related wikipedia page using the selected method
}
