% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_wikipedia_page.R
\name{search_google_for_wikipedia_page}
\alias{search_google_for_wikipedia_page}
\title{Searches Google for Wikipedia Page}
\usage{
search_google_for_wikipedia_page(search_string)
}
\arguments{
\item{search_string}{The string to be searched on google. Should be a character}
}
\value{
A dataframe consisting of the search result and the associated page_name
}
\description{
This function takes a location string and finds it's related wikipedia page using google search and rvest.
}
\details{
This function may have some issues.It relies on a specific formatting of google search results that may change in the future to work.
Use with some caution.
Also: may have to account for simple.wiki
}
